JV: Add a more sophisticated default basic expression checker
Previously the default was not to classify anything as an advanced-only expression, however that means out of the box the editor will explode on things that it doesn't support. By providing a sane default, there's less of a chance of that, but it still gives the user the option to completely override it if they build additional support in their theme for something the default theme doesn't support, or to provide additional restrictions on top of the new default check.